:truck: refactor(tests): Rename helper files.
