+++ imports/api/_dev/populate/uploads.ts
@@ -1,7 +1,8 @@
-import {randomPNGBuffer, randomPNGDataURI} from '../test/png';
-import {Uploads} from './uploads';
+import {FileRef} from 'meteor/ostrio:files';
+import {randomPNGBuffer, randomPNGDataURI} from '../../../test/png';
+import {MetadataType, Uploads} from '../../uploads';
 
-export const newUpload = async (invocation) => {
+export const newUpload = async (invocation): Promise<FileRef<MetadataType>> => {
 	const type = 'image/png';
 	const fileName = 'pic.png';
 	if (Meteor.isServer) {
@@ -27,26 +28,18 @@ export const newUpload = async (invocation) => {
 	}
 
 	return new Promise((resolve, reject) => {
-		const upload = Uploads.insert(
-			{
-				file: randomPNGDataURI(),
-				isBase64: true,
-				fileName,
-				chunkSize: 'dynamic',
+		Uploads.insert({
+			file: randomPNGDataURI(),
+			isBase64: true,
+			fileName,
+			chunkSize: 'dynamic',
+			onUploaded: (error, fileRef) => {
+				if (error) {
+					reject(error);
+				} else {
+					resolve(fileRef);
+				}
 			},
-			false,
-		);
-
-		upload.on('end', (err, fileObject) => {
-			if (err) {
-				reject(err);
-			} else {
-				resolve(fileObject);
-			}
 		});
-
-		upload.start();
 	});
 };
-
-export {Uploads} from './uploads';
