--- imports/api/collection/attachments.mock.ts
@@ -1,1 +1,0 @@
-export {Attachments} from './attachments';
