--- imports/api/collection/availability.mock.ts
@@ -1,1 +1,0 @@
-export {Availability} from './availability';
