--- imports/api/collection/books.mock.ts
@@ -1,1 +1,0 @@
-export {Books} from './books';
