+++ imports/api/endpoint/appointments/schedule.tests.ts
@@ -4,11 +4,15 @@ import {assert} from 'chai';
 
 import {Random} from 'meteor/random';
 
-import {Appointments, newAppointment} from '../../collection/appointments.mock';
-import {Availability} from '../../collection/availability.mock';
-import {slot} from '../../availability';
-import {beginningOfTime, endOfTime} from '../../../util/datetime';
 import {dropIds, server} from '../../../test/fixtures';
+
+import {beginningOfTime, endOfTime} from '../../../util/datetime';
+
+import {Appointments} from '../../collection/appointments';
+import {Availability} from '../../collection/availability';
+import {newAppointment} from '../../_dev/populate/appointments';
+
+import {slot} from '../../availability';
 
 server(__filename, () => {
 	it('fills availability', async () => {
